Move builder tests into separate file
Moved the tests associated with the builder module into their own file for
readability reasons. The builder implementation plus the tests was getting a bit
ugly.
